* Added quick check for connection vars * Fixed SPARQL testing * Moved cpt tests inside trecov.sh

+++ binsrc/tests/suite/testall.sh
@@ -211,7 +211,7 @@ rm -f audit.txt core debug.txt
 #
 ./tvsp.sh
 #./tupgrade_recov.sh
-./tcpt.sh
+# ./tcpt.sh (inside trecov.sh)
 ./trecov.sh
 ./trecov_schema.sh
 ./tsql.sh
@@ -247,6 +247,7 @@ fi
 ./bpel.sh
 ./tdav_meta.sh
 TEST_SPARQL=yes ./tpcd.sh
+./tsparql.sh
 
 #XXX: not tested yet on Win32
 if [ "x$HOST_OS" != "x" ]
@@ -495,21 +496,27 @@ then
     exit 3
 fi
 #fi
+if test \! -f tsparql.output
+then
+    ECHO "***ABORTED: No tsparql.output"
+    exit 3
+fi
+
 #
 #  Check if the tests logged any failures
 #
-RUN egrep '"\*\*\*.*FAILED|\*\*\*.*ABORTED"' tvsp.output trecov.output tsql.output tsql2.output tsql3.output tsec.output rtest.output gtkbench.output thttp.output tproxy.output tdav.output twcopy.output $sqlo_outputs timsg.output tvad.output trepl.output nwxml.output txslt.output obackup.output tjdbc.output inprocess.output tvspxex.output tsoap12.output trecov_schema.output msdtc.output ttutorial.output bpel.output tdav_meta.output tpcd.output tlubm.output tcpt.output # tupgrade_recov.output 
+RUN egrep '"^(\*\*\*.*FAILED|\*\*\*.*ABORTED)"' tvsp.output trecov.output tsql.output tsql2.output tsql3.output tsec.output rtest.output gtkbench.output thttp.output tproxy.output tdav.output twcopy.output $sqlo_outputs timsg.output tvad.output trepl.output nwxml.output txslt.output obackup.output tjdbc.output inprocess.output tvspxex.output tsoap12.output trecov_schema.output msdtc.output ttutorial.output bpel.output tdav_meta.output tpcd.output tlubm.output tcpt.output tsparql.output # tupgrade_recov.output 
 if test $STATUS -eq 0
 then
     ECHO ""
     LINE
     ECHO "=  WARNING: Some tests failed. See *.output in this directory" `pwd`
-    egrep '\*\*\*.*FAILED|\*\*\*.*ABORTED' tvsp.output trecov.output tsql.output tsql2.output tsql3.output tsec.output rtest.output gtkbench.output thttp.output tproxy.output tdav.output twcopy.output $sqlo_outputs timsg.output tvad.output trepl.output nwxml.output txslt.output obackup.output tjdbc.output inprocess.output tvspxex.output tsoap12.output trecov_schema.output msdtc.output ttutorial.output bpel.output tdav_meta.output tpcd.output tlubm.output tcpt.output # tupgrade_recov.output 
+    egrep '^(\*\*\*.*FAILED|\*\*\*.*ABORTED)' tvsp.output trecov.output tsql.output tsql2.output tsql3.output tsec.output rtest.output gtkbench.output thttp.output tproxy.output tdav.output twcopy.output $sqlo_outputs timsg.output tvad.output trepl.output nwxml.output txslt.output obackup.output tjdbc.output inprocess.output tvspxex.output tsoap12.output trecov_schema.output msdtc.output ttutorial.output bpel.output tdav_meta.output tpcd.output tlubm.output tcpt.output tsparql.output # tupgrade_recov.output 
     LINE
     rm -f audit.txt
  
     #grail error
-    RUN egrep '"\*\*\*.*FAILED|\*\*\*.*ABORTED"' trecov.output 
+    RUN egrep '"^(\*\*\*.*FAILED|\*\*\*.*ABORTED)"' trecov.output 
     if test $STATUS -eq 0
     then
 	rm -rf grail_backup
@@ -518,7 +525,7 @@ then
 	cp trecov.output grail_backup/trecov.out.err
     fi	
     #grail error part2
-    RUN egrep '"\*\*\*.*FAILED|\*\*\*.*ABORTED"' trecov_schema.output 
+    RUN egrep '"^(\*\*\*.*FAILED|\*\*\*.*ABORTED)"' trecov_schema.output 
     if test $STATUS -eq 0
     then
 	rm -rf grail_backup2
